% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EB_estimators.R
\name{ebest}
\alias{ebest}
\title{Empirical Bayes Estimators for Population-Specific Association}
\usage{
ebest(K, betas, ses)
}
\arguments{
\item{K}{Number of populations}

\item{betas}{Vector of regression coefficient estimates for the K populations, the first one is considered as the population of interest}

\item{ses}{Vector of standard errors of the corresponding regression coefficient estimates for the \emph{K} populations}
}
\description{
ebest provides joint and pairwise empirical Bayes (EB) estimators for population-specific association analysis,
while leveraging information from other populations, accounting for both the uncertainty of the estimates and
heterogeneity across populations. ebest also provides the conventional meta-analysis fixed effect estimator
for combining the estimates across populations.
}
\details{
Estimates and standard errors for the meta-analysis estimator (b.meta, se.meta), joint EB estimator (b.joint, se.joint),
and pairwise EB estimator (b.pairwise, se.pairwise).
If (K = 2), the joint and pairwise EB estimators are same and only the joint EB estimator is outputed.
}
\note{
In the example give below, 3 populations are used (K value), and the 3 values for the population sizes, p.freq, and betas each given in the 3 vectors are used to calculate the parameters "beta" and "ses" in the function ebest.
}
\examples{

K = 3
est = c(0.1, 0.2, 0.3)
se = c(0.2, 0.2, 0.2)
ebest(K = K, betas = est, ses = se)

}
\references{
Hsu, L., Kooperberg, A., Reiner, A.P. and Kooperberg, C., (2023) An empirical Bayes approach to improving population‐specific genetic association estimation by leveraging cross‐population data. Genetic Epidemiology, 47(1), pp.45-60.
}
\author{
Contributors: Li Hsu, Lilian Law
}
